bug 365 fixed: categories.rank was limited to 255 due to storage limitation. Maximum is 65535 now (should be good for a long time).
git-svn-id: http://piwigo.org/svn/branches/branch-1_6@1311 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ install/phpwebgallery_structure.sql
@@ -28,7 +28,7 @@ CREATE TABLE `phpwebgallery_categories` (
   `id_uppercat` smallint(5) unsigned default NULL,
   `comment` text,
   `dir` varchar(255) default NULL,
-  `rank` tinyint(3) unsigned default NULL,
+  `rank` smallint(5) unsigned default NULL,
   `status` enum('public','private') NOT NULL default 'public',
   `site_id` tinyint(4) unsigned default '1',
   `visible` enum('true','false') NOT NULL default 'true',
